modified :  - delete user 확인 강화

+++ src/libs/firebase/auth.ts
@@ -53,6 +53,7 @@ export const logout = () =>{
 }
 
 export const withDrawlUser = async () =>{
+    if(auth.currentUser?.isAnonymous) return
     try{
         drawlUserDeleteSettingData(auth.currentUser?.email as string);
         return await auth.currentUser?.delete();
